New modules (#29)
* new implementation and testing folder

* removed example folder, react-scripts caused dependecy issues

* added new implementations and utils functions, 1-file structure right now since module doesn't work with Puppeteer eval

* added method to serialize form, added function to scrub the changes global variable to remove circular references

* refactored TreeNode class to its own module

* modified toSerializable method, disabled some eslint rules

* added function to serialize memoizedProps

* added a stretch feature starter code for time-travel

* placeholder type for now
